chore: move OutputBuiltinRunner to own file
